refactor: move all destination services hook into dedicated file
