--- src/services/connector/ConnectorType.ts
@@ -1,9 +1,0 @@
-export type Connector = {
-  description: string;
-  // configuration: {},
-  tombstone: boolean;
-  user: string;
-  org: string;
-  create_time: string;
-  update_time: string;
-};
