+++ src/services/connector/destination/index.ts
@@ -1,0 +1,15 @@
+import useDestination from "./useDestination";
+import useDestinations from "./useDestinations";
+import useDestinationWithPipelines from "./useDestinationWithPipelines";
+import useDestinationsWithPipelines from "./useDestinationsWithPipelines";
+import useDestinationDefinitions from "./useDestinationDefinitions";
+import useCreateDestination from "./useCreateDestination";
+
+export {
+  useDestination,
+  useDestinations,
+  useDestinationWithPipelines,
+  useDestinationsWithPipelines,
+  useDestinationDefinitions,
+  useCreateDestination,
+};
