+++ src/services/connector/index.ts
@@ -1,2 +1,1 @@
-import DestinationTable from "./DestinationTable";
-import type { DestinationTableProps } from "./DestinationTable";
+export * from "./destination";
